// app/components/reactive-forms/interface.ts
export interface Question1 {
  value: string;
  viewValue: string;
}

export interface Question2 {
  value: string;
  viewValue: string;
}

export interface Question3 {
  value: string;
  viewValue: string;

}

export interface Question4 {
  value: string;
  viewValue: string;
}